+++ pkg/admin/controller.go
@@ -6,6 +6,8 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/pyroscope-io/pyroscope/pkg/server/httputils"
+	pstorage "github.com/pyroscope-io/pyroscope/pkg/storage"
 	"net/http"
 
 	"github.com/gorilla/mux"
@@ -14,10 +16,16 @@ import (
 	"github.com/pyroscope-io/pyroscope/pkg/model"
 )
 
-type Controller struct {
-	log *logrus.Logger
+type Storage interface {
+	pstorage.AppNameGetter
+	pstorage.AppGetter
+	pstorage.AppDeleter
+}
 
-	adminService   *AdminService
+type Controller struct {
+	log            *logrus.Logger
+	httpUtils      httputils.Utils
+	storage        Storage
 	userService    UserService
 	storageService StorageService
 }
@@ -32,50 +40,16 @@ type StorageService interface {
 
 func NewController(
 	log *logrus.Logger,
-	adminService *AdminService,
+	storage Storage,
 	userService UserService,
 	storageService StorageService) *Controller {
 	return &Controller{
 		log: log,
 
-		adminService:   adminService,
+		storage:        storage,
 		userService:    userService,
 		storageService: storageService,
 	}
-}
-
-// HandleGetApps handles GET requests
-func (ctrl *Controller) HandleGetApps(w http.ResponseWriter, _ *http.Request) {
-	appNames := ctrl.adminService.GetApps()
-
-	w.WriteHeader(http.StatusOK)
-	ctrl.writeResponseJSON(w, appNames)
-}
-
-type DeleteAppInput struct {
-	Name string `json:"name"`
-}
-
-// HandleDeleteApp handles DELETE requests
-func (ctrl *Controller) HandleDeleteApp(w http.ResponseWriter, r *http.Request) {
-	var payload DeleteAppInput
-
-	err := json.NewDecoder(r.Body).Decode(&payload)
-	if err != nil {
-		ctrl.writeError(w, http.StatusBadRequest, err, "")
-		return
-	}
-
-	err = ctrl.adminService.DeleteApp(payload.Name)
-	if err != nil {
-		// TODO how to distinguish
-		// it was a bad request
-		// or an internal server error
-		ctrl.writeError(w, http.StatusInternalServerError, err, "")
-		return
-	}
-
-	w.WriteHeader(http.StatusOK)
 }
 
 type UpdateUserRequest struct {