+++ webapp/javascript/components/TimelineChart/TimelineChartWrapper.tsx
@@ -2,11 +2,19 @@
 /* eslint-disable react/no-did-update-set-state */
 /* eslint-disable react/destructuring-assignment */
 import React, { ReactNode } from 'react';
-import { Timeline } from '@webapp/models/timeline';
 import Color from 'color';
+import type { Group } from '@pyroscope/models/src';
+import type { Timeline } from '@webapp/models/timeline';
 import { formatAsOBject } from '@webapp/util/formatDate';
+import Legend from '@webapp/pages/tagExplorer/components/Legend';
 import TimelineChart from './TimelineChart';
 import styles from './TimelineChartWrapper.module.css';
+
+export interface TimelineGroupData {
+  data: Group;
+  tagName: string;
+  color?: Color;
+}
 
 interface TimelineData {
   data?: Timeline;
@@ -20,7 +28,29 @@ interface Marking {
   overlayColor?: string | Color;
 }
 
-type TimelineChartWrapperProps = {
+type TimelineDataProps =
+  | {
+      /** timelineA refers to the first (and maybe unique) timeline */
+      timelineA: TimelineData;
+      /** timelineB refers to the second timeline, useful for comparison view */
+      timelineB?: TimelineData;
+      /** to manage strict data type */
+      mode: 'singles';
+    }
+  | {
+      /** timelineGroups refers to group of timelines, useful for explore view */
+      timelineGroups: TimelineGroupData[];
+      /** if there is active group, the other groups should "dim" themselves */
+      activeGroup: string;
+      /** show or hide tags legend, useful forr disabling single timeline legend */
+      showTagsLegend: boolean;
+      /** to set active tagValue using <Legend /> */
+      handleGroupByTagValueChange: (groupByTagValue: string) => void;
+      /** to manage strict data type */
+      mode: 'multiple';
+    };
+
+type TimelineChartWrapperProps = TimelineDataProps & {
   /** the id attribute of the element float will use to apply to, it should be unique */
   id: string;
 
@@ -28,10 +58,6 @@ type TimelineChartWrapperProps = {
   onSelect: (from: string, until: string) => void;
   format: 'lines' | 'bars';
 
-  /** timelineA refers to the first (and maybe unique) timeline */
-  timelineA: TimelineData;
-  /** timelineB refers to the second timeline, useful for comparison view */
-  timelineB?: TimelineData;
   height?: string;
 
   /** refers to the highlighted selection */
@@ -55,6 +81,7 @@ class TimelineChartWrapper extends React.Component<
   // eslint-disable-next-line react/static-property-placement
   static defaultProps = {
     format: 'bars',
+    mode: 'singles',
   };
 
   constructor(props: TimelineChartWrapperProps) {
@@ -199,7 +226,59 @@ class TimelineChartWrapper extends React.Component<
 
   render = () => {
     const { flotOptions } = this.state;
+
+    if (this.props.mode === 'multiple') {
+      const { timelineGroups, activeGroup, showTagsLegend, id, timezone } =
+        this.props;
+
+      const customFlotOptions = {
+        ...flotOptions,
+        xaxis: {
+          ...flotOptions.xaxis,
+          autoscaleMargin: null,
+          timezone: timezone || 'browser',
+        },
+      };
+
+      const centeredTimelineGroups =
+        timelineGroups &&
+        timelineGroups.map(({ data, color, tagName }) => {
+          return {
+            data: centerTimelineData({ data }),
+            color:
+              activeGroup && activeGroup !== tagName
+                ? color?.fade(0.75)
+                : color,
+          };
+        });
+
+      return (
+        <>
+          <TimelineChart
+            onSelect={this.props.onSelect}
+            className={styles.wrapper}
+            // eslint-disable-next-line react/destructuring-assignment
+            data-testid={this.props['data-testid']}
+            id={id}
+            options={customFlotOptions}
+            data={centeredTimelineGroups}
+            width="100%"
+            height={this.props.height || '100px'}
+          />
+          {showTagsLegend && (
+            <Legend
+              groups={timelineGroups}
+              handleGroupByTagValueChange={
+                this.props.handleGroupByTagValueChange
+              }
+            />
+          )}
+        </>
+      );
+    }
+
     const { id, timelineA, timezone, title } = this.props;
+
     // TODO deep copy
     let timelineB = this.props.timelineB
       ? JSON.parse(JSON.stringify(this.props.timelineB))
@@ -212,7 +291,7 @@ class TimelineChartWrapper extends React.Component<
         // In case there are few chunks left, then we'd like to add some margins to
         // both sides making it look more centers
         autoscaleMargin:
-          timelineA.data && timelineA.data.samples.length > 3 ? null : 0.005,
+          timelineA?.data && timelineA.data.samples.length > 3 ? null : 0.005,
         timezone: timezone || 'browser',
       },
     };
@@ -263,7 +342,6 @@ class TimelineChartWrapper extends React.Component<
           id={id}
           options={customFlotOptions}
           data={data}
-          //        data={d}
           width="100%"
           height={this.props.height || '100px'}
         />